+++ src/Symfony/Bundle/TwigBundle/Tests/Functional/EmptyAppTest.php
@@ -15,7 +15,7 @@ use Symfony\Bundle\TwigBundle\Tests\TestCase;
 use Symfony\Bundle\TwigBundle\TwigBundle;
 use Symfony\Component\Config\Loader\LoaderInterface;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
-use Symfony\Component\ErrorRenderer\ErrorRenderer;
+use Symfony\Component\ErrorHandler\ErrorRenderer\HtmlErrorRenderer;
 use Symfony\Component\Filesystem\Filesystem;
 use Symfony\Component\HttpKernel\Kernel;
 
@@ -61,7 +61,8 @@ class EmptyAppKernel extends Kernel
     public function registerContainerConfiguration(LoaderInterface $loader)
     {
         $loader->load(static function (ContainerBuilder $container) {
-            $container->register('error_renderer', ErrorRenderer::class);
+            $container->register('error_renderer.html', HtmlErrorRenderer::class);
+            $container->setAlias('error_renderer', 'error_renderer.html');
             $container->setParameter('debug.file_link_format', null);
         });
     }