+++ src/Symfony/Bridge/Twig/CHANGELOG.md
@@ -12,6 +12,7 @@ CHANGELOG
 4.4.0
 -----
 
+ * added a new `TwigErrorRenderer` for `html` format, integrated with the `ErrorHandler` component
  * marked all classes extending twig as `@final`
  * deprecated to pass `$rootDir` and `$fileLinkFormatter` as 5th and 6th argument respectively to the 
    `DebugCommand::__construct()` method, swap the variables position.